Bookingnumber on save, remove booking from Admin bookings
Implemented so that a Booking is removed by a identifier string.

The string in question is a pseudo-random subset of letters and numbers - and relies on the implementation of toString of floating point numbers.

This pseudo-randomness is not sufficient to account for randomness in the absolute - nor does it guarantee uniqueness - But the expectancy of a collision on the numeral basis of combinations of letters and numbers is not estimated to be great - as the upper boundary would be 36 (26 letters + 10 numerals) to the power of 10. (So, 36^10)

Other than that, implementation stands to be that a booking is saved with a bookingNumber (numerals and letters) - and that it's spliced away from respective loaded JSON file, and then saved back as the updated version.