Added numbered seats taken to Modal and fixed a bug
There was a potential bug in terms of a modal not being closed and just being "layered", meaning that many occurred bookings would have darkened the pages background if the modal was not closed and more bookings occurred.

To fix this, simply checked that the modal has not been rendered yet and that there are seats which had a overlap.