Fixed bug where Ticket info rendered incorrectly
There was a bug that was when you overwrote the last booked ticket for another person, there was nothing that actually cleared out the HTML and reset the state of the HTML elements to be as if no seat was taken.

The functionality worked under the pre-tense of there being seats still chosen, but not when there was no seats left.

Fixed this issue.