Implemented so that liveBooking does not overlap in bookings
There was a bug that caused an overlap in terms of booked Seats - Where if a used booked seats that another one had marked as wanting them, it would still register as being that the user when going to book - would get seats booked that were meant to be disabled or already taken.

Resolved this issue.

Had to work around Splice once more, because it only runs once and then stops execution - so worked around it with slice and concat.

Updated SessionStorage of the selectedShow with the info as well, to resolve the issue of the bug with the act of booking.