Fixed manual href Abuse
If you were to manually type in the hrefs, you could access different parts of the site that you are not meant to be able to access - Fixed this by just making a check against the sessionStorage if you are logged in or not - in regards to loginPage and registerPage.

If you are logged in, you are put to your userPage, for both cases.

If you are NOT logged in, you are directed to the correct place.