Added Modal and prevented Booking of past shows
Added so that if a date is attempted to be picked, that is in the past, it will pop up a Modal and not render those movies - keeps track of the previous date when clicking too, as to know what value to default back to when not re-rendering with the faulty past in-time value.

Wanted to implement disabled attributes for previous dates to make it extra clear, but would require additional imports such as jQuery UI and additional efforts in re-constructing the datePicker - which was deemed as too tight on time to integrate at the moment.