Corrected bug in CSS that caused Min-width to override max-width
In CSS - min-width with a declaration of max-content, will override the priority of max-width - if both are present.

This caused an interaction where min-width would cause Firefox browsers and Safari to understand the failed register Modal as being a sideways scrollable element, which it is not intended to be.

Fixed this by removing the normal circumstanced min-width of max-content, because media queries handles the cases of where it is needed to be stated otherwise - in terms of widths and sizings.